Move lobby login challenge/response key constants to login subpackage
