+++ game-core/src/main/java/games/strategy/engine/lobby/common/login/LobbyLoginChallengeKeys.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.lobby.common;
+package games.strategy.engine.lobby.common.login;
 
 /**
  * The property keys that may be present in a lobby authentication protocol challenge.
