+++ lobby/src/main/java/games/strategy/engine/lobby/server/login/LobbyLoginValidator.java
@@ -17,8 +17,8 @@ import com.google.common.base.Strings;
 
 import games.strategy.engine.config.lobby.LobbyPropertyReader;
 import games.strategy.engine.lobby.common.LobbyConstants;
-import games.strategy.engine.lobby.common.LobbyLoginChallengeKeys;
-import games.strategy.engine.lobby.common.LobbyLoginResponseKeys;
+import games.strategy.engine.lobby.common.login.LobbyLoginChallengeKeys;
+import games.strategy.engine.lobby.common.login.LobbyLoginResponseKeys;
 import games.strategy.engine.lobby.server.User;
 import games.strategy.engine.lobby.server.db.AccessLogController;
 import games.strategy.engine.lobby.server.db.BadWordController;
