+++ lobby/src/test/java/games/strategy/engine/lobby/server/login/LobbyLoginValidatorTest.java
@@ -31,8 +31,8 @@ import com.google.common.collect.ImmutableMap;
 import games.strategy.engine.config.MemoryPropertyReader;
 import games.strategy.engine.config.lobby.LobbyPropertyReader;
 import games.strategy.engine.lobby.common.LobbyConstants;
-import games.strategy.engine.lobby.common.LobbyLoginChallengeKeys;
-import games.strategy.engine.lobby.common.LobbyLoginResponseKeys;
+import games.strategy.engine.lobby.common.login.LobbyLoginChallengeKeys;
+import games.strategy.engine.lobby.common.login.LobbyLoginResponseKeys;
 import games.strategy.engine.lobby.server.TestUserUtils;
 import games.strategy.engine.lobby.server.User;
 import games.strategy.engine.lobby.server.db.BadWordDao;
